chore: rename module to jaeger-influxdb
