+++ jaeger-influxdb/internal/common.go
